Revert "feat: Implementation of the Badge component Style API (#3650)"
This reverts commit afdbe00651ba2d3eb2436229e882a63c2335623e.
