+++ src/badge/interfaces.ts
@@ -23,32 +23,4 @@ export interface BadgeProps extends BaseComponentProps {
    * Text displayed inside the badge.
    */
   children?: React.ReactNode;
-
-  /**
-   * Specifies an object of selectors and properties that are used to apply custom styles.
-   *
-   * - `root.background` (string) - (Optional) Background for badge.
-   * - `root.borderColor` (string) - (Optional) Border color for badge.
-   * - `root.borderRadius` (string) - (Optional) Border radius style.
-   * - `root.borderWidth` (string) - (Optional) Border width style.
-   * - `root.color` (string) - (Optional) Text color for badge.
-   * - `root.paddingBlock` (string) - (Optional) Block dimension padding.
-   * - `root.paddingInline` (string) - (Optional) Inline dimension padding.
-   * @awsuiSystem core
-   */
-  style?: BadgeProps.Style;
-}
-
-export namespace BadgeProps {
-  export interface Style {
-    root?: {
-      background?: string;
-      borderColor?: string;
-      borderRadius?: string;
-      borderWidth?: string;
-      color?: string;
-      paddingBlock?: string;
-      paddingInline?: string;
-    };
-  }
 }
