+++ src/badge/styles.scss
@@ -16,8 +16,6 @@
   border-start-end-radius: awsui.$border-radius-badge;
   border-end-start-radius: awsui.$border-radius-badge;
   border-end-end-radius: awsui.$border-radius-badge;
-  border-block-style: solid;
-  border-inline-style: solid;
   padding-block: 0;
   padding-inline: awsui.$space-xs;
   color: awsui.$color-text-notification-default;
